refactor: made separate adminTableRow component from adminTableBody
