+++ src/components/admin/AdminTableBody.tsx
@@ -3,32 +3,12 @@
 import React, { useState, useEffect } from "react";
 import { fetchTestimonials } from "@/app/hooks/admin/fetchTestimonials";
 import { TestimonialAdminData } from "@/types/testimonials.type";
-import StatusColumnButtons from "./StatusColumnButtons";
-import TableCellDropDown from "./TableCellDropDown";
 import { supabase } from "@/lib/supabaseClient";
-
-const columns = [
-  { header: "Name", accessor: "name" },
-  { header: "Email", accessor: "email" },
-  { header: "Review", accessor: "testimonial" },
-  { header: "Rating", accessor: "rating" },
-  { header: "Status", accessor: "is_approved" },
-];
-
-const maxLengths: Record<string, number> = {
-  testimonial: 35,
-  name: 20,
-  email: 30,
-};
+import AdminTableRow from "./AdminTableRow";
 
 const AdminTableBody: React.FC = () => {
   const [error, setError] = useState("");
   const [testimonials, setTestimonials] = useState<TestimonialAdminData[]>([]);
-  // const [statuses, setStatuses] = useState<Record<number, string>>({});
-
-  // const handleStatusChange = (rowIdx: number, status: string) => {
-  //     setStatuses((prev) => ({ ...prev, [rowIdx]: status }));
-  // };
 
   useEffect(() => {
     const loadData = async () => {
@@ -83,38 +63,7 @@ const AdminTableBody: React.FC = () => {
   return (
     <tbody className="text-center text-sm w-full">
       {testimonials.map((row, rowIdx) => (
-        <tr
-        key={rowIdx}
-        className={`${rowIdx === testimonials.length - 1 ? "" : "border-b-neutral-300 border-b-1"}`}
-        >
-          {columns.map((col, colIdx) => {
-            const cellValue = row[col.accessor as keyof TestimonialAdminData];
-            const maxLength = maxLengths[col.accessor] ?? Infinity;
-
-            const shouldTruncate =
-              typeof cellValue === "string" && cellValue.length > maxLength;
-
-            const shortText = shouldTruncate
-              ? `${cellValue.slice(0, maxLength)}...`
-              : cellValue;
-
-              return (
-                <td key={colIdx} className="px-4 py-2 text-center h-14">
-                  { col.accessor === "is_approved" ? (
-                    row.is_approved === "Accepted" || row.is_approved === "Declined" ? (
-                      <span>{row.is_approved}</span>
-                    ) : (
-                      <StatusColumnButtons rowId={row.testimonial_id} />
-                    )
-                  ) : shouldTruncate ? (
-                    <TableCellDropDown shortText={String(shortText)} fullText={cellValue} isReview={col.accessor === "testimonial"}/>
-                  ) : (
-                    cellValue
-                  )}
-                </td>
-              );
-          })}
-        </tr>
+        <AdminTableRow key={row.testimonial_id} row={row} isLastRow={rowIdx === testimonials.length - 1} />
       ))
       }
     </tbody>
@@ -122,20 +71,3 @@ const AdminTableBody: React.FC = () => {
 };
 
 export default AdminTableBody;
-
-
-// return (
-//   <td key={colIdx} className="px-4 py-2 text-center h-14">
-//     { col.accessor === "is_approved" ? (
-//       statuses[rowIdx] ? (
-//         <span>{statuses[rowIdx]}</span>
-//       ) : (
-//         <StatusColumnButtons rowId={row.id} onStatusUpdate={(status) => handleStatusChange(rowIdx, status)}/>
-//       )
-//     ) : shouldTruncate ? (
-//       <TableCellDropDown shortText={String(shortText)} fullText={cellValue} isReview={col.accessor === "testimonial"}/>
-//     ) : (
-//       cellValue
-//     )}
-//   </td>
-// );
